+++ templates/partials/sidebar/trackdetails.php
@@ -25,7 +25,7 @@
 			<dl class="tags">
 				<dt ng-repeat-start="tag in details.tags | orderBy:[tagRank, 'key'] | filter:tagAlwaysShown">{{ formatDetailName(tag.key) }}</dt>
 				<dd ng-repeat-end ng-class="{clickable: tagHasDetails(tag)}" ng-click="showTagDetails(tag)"
-				>{{ formatDetailValue(tag.value, tag.key) }}<button class="icon-info" ng-if="tagHasDetails(tag)"></button></dd>
+				><span ng-bind-html="formatDetailValue(tag.value, tag.key)"></span><button class="icon-info" ng-if="tagHasDetails(tag)"></button></dd>
 	
 				<dt ng-if="details.length">length</dt>
 				<dd ng-if="details.length">{{ details.length | playTime }}</dd>
@@ -33,8 +33,8 @@
 			<p class="show-more-less" ng-if-start="anyCollapsibleTags(details.tags)" ng-init="collapsed = true" ng-click="collapsed = false" ng-show="collapsed" translate>Show more…</p>
 			<dl class="tags" ng-show="!collapsed">
 				<dt ng-repeat-start="tag in details.tags | orderBy:[tagRank, 'key'] | filter:tagShownWhenExpanded">{{ formatDetailName(tag.key) }}</dt>
-				<dd ng-repeat-end ng-class="{clickable: tagHasDetails(tag)}" ng-click="showTagDetails(tag)" ng-bind-html="formatDetailValue(tag.value, tag.key)"
-				><button class="icon-info" ng-if="tagHasDetails(tag)"></button></dd>
+				<dd ng-repeat-end ng-class="{clickable: tagHasDetails(tag)}" ng-click="showTagDetails(tag)"
+				><span ng-bind-html="formatDetailValue(tag.value, tag.key)"></span><button class="icon-info" ng-if="tagHasDetails(tag)"></button></dd>
 			</dl>
 			<p class="show-more-less" ng-if-end ng-show="!collapsed" ng-click="collapsed = true" translate>Show less…</p>
 		</div>