+++ CHANGELOG.md
@@ -26,8 +26,9 @@
 - Track details pane:
   * Collapse the more exotic tags by default and show them by clicking "Show more…"
   * Make all MusicBrainz Ids from the tags into links to the MusicBrainz site (like previously done with MbIds from Last.fm)
-  * On tag names, abbreaviate "musicbrainz" as "mb" to not truncate the more important parts of the name
+  * On tag names, abbreviate "musicbrainz" as "mb" to not truncate the more important parts of the name
   * Show also MusicBrainz Recording Id stored in id3v2.4 tag `UFID` (MusicBrainz Picard uses this on mp3 files)
+  * Show all values of multi-valued metadata tags
 
 ### Fixed
 - Deprecation warnings printed on PHP 8.3+ while executing the Music background tasks